# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "User Authentication API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "summary": "User Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Username",
                    "example": "johndoe"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6,
                    "description": "User password",
                    "example": "password123"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "Authentication token",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  },
                  "required": [
                    "token"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "User Registration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Username",
                    "example": "johndoe"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User email",
                    "example": "john@example.com"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6,
                    "description": "User password",
                    "example": "password123"
                  }
                },
                "required": [
                  "username",
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "$ref": "#/components/schemas/userId"
                    },
                    "username": {
                      "type": "string",
                      "description": "Username",
                      "example": "johndoe"
                    },
                    "email": {
                      "type": "string",
                      "format": "email",
                      "description": "User email",
                      "example": "john@example.com"
                    }
                  },
                  "required": [
                    "userId",
                    "username",
                    "email"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/user": {
      "get": {
        "summary": "Get User Details",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "User details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "$ref": "#/components/schemas/userId"
                    },
                    "username": {
                      "type": "string",
                      "description": "Username",
                      "example": "johndoe"
                    },
                    "email": {
                      "type": "string",
                      "format": "email",
                      "description": "User email",
                      "example": "john@example.com"
                    }
                  },
                  "required": [
                    "userId",
                    "username",
                    "email"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "userId": {
        "type": "string",
        "description": "A unique identifier for a user",
        "example": "user123"
      }
    }
  }
}